figured out how to get postgres-operator into its own namespace
problem was launching pods with the RBAC configuration.  kept saying
it coudn't list operatorconfigurations as the service account user.

turns out that in order to list operatorconfigurations you need to
have permissions as Cluster scope.  so I created the service account
in its own namespace but I gave it a ClusterRole and ClusterRoleBinding